replace static_cast by dynamic_cast

+++ src/Calculator/Views/HistoryList.xaml.cpp
@@ -42,11 +42,11 @@ HistoryList::HistoryList()
 
 void HistoryList::ListView_ItemClick(_In_ Object^ sender, _In_ ItemClickEventArgs^ e)
 {
-    HistoryViewModel^ historyVM = static_cast<HistoryViewModel^>(this->DataContext);
-    HistoryItemViewModel^ clickedItem = safe_cast<HistoryItemViewModel^>(e->ClickedItem);
+    HistoryViewModel^ historyVM = dynamic_cast<HistoryViewModel^>(this->DataContext);
+    HistoryItemViewModel^ clickedItem = dynamic_cast<HistoryItemViewModel^>(e->ClickedItem);
 
     // When the user clears the history list in the overlay view and presses enter, the clickedItem is nullptr
-    if (clickedItem != nullptr)
+    if (clickedItem != nullptr && historyVM != nullptr)
     {
         historyVM->ShowItem(clickedItem);
     }
@@ -55,8 +55,7 @@ void HistoryList::ListView_ItemClick(_In_ Object^ sender, _In_ ItemClickEventArg
 void HistoryList::OnDeleteMenuItemClicked(_In_ Object^ sender, _In_ RoutedEventArgs^ e)
 {
     auto listViewItem = HistoryContextMenu->Target;
-    auto itemViewModel = safe_cast<HistoryItemViewModel^>(HistoryListView->ItemFromContainer(listViewItem));
-
+    auto itemViewModel = dynamic_cast<HistoryItemViewModel^>(HistoryListView->ItemFromContainer(listViewItem));
     if (itemViewModel != nullptr)
     {
         Model->DeleteItem(itemViewModel);
@@ -65,9 +64,11 @@ void HistoryList::OnDeleteMenuItemClicked(_In_ Object^ sender, _In_ RoutedEventA
 
 void HistoryList::OnDeleteSwipeInvoked(_In_ MUXC::SwipeItem^ sender, _In_ MUXC::SwipeItemInvokedEventArgs^ e)
 {
-    auto swipedItem = safe_cast<HistoryItemViewModel^>(e->SwipeControl->DataContext);
-
-    Model->DeleteItem(swipedItem);
+    auto swipedItem = dynamic_cast<HistoryItemViewModel^>(e->SwipeControl->DataContext);
+    if (swipedItem != nullptr)
+    {
+        Model->DeleteItem(swipedItem);
+    }
 }
 
 void HistoryList::ScrollToBottom()
